Adding unions to convert a primitive data type to bytes while producing 0 garbage.
ADD: Added union types for several of the basic primitives that I intend to serialize with NetStream
ADD: Added some tests to prove the functionality of the unions is infact working.

NOTE: Why not use BitConverter? Because it creates unnecessary garbage. Using structs + StructLayout + LayoutKind.Explicit may not be supported on every platform and endianess is not guaranteed to be correct (little endian) for now.